Fixed ambience/reflectance calculations so they work correctly

+++ src/me/jessicaalohse/raytracer/shapes/Surface.java
@@ -8,7 +8,7 @@ public interface Surface {
 	public double getT();
 	public boolean hit(Ray ray, double tSubZero, double tSub1, float time);
 	public RGB getColor();
-	public double getReflectance();
+	public float getReflectance();
 	public boolean shadowHit(Ray ray, float tSubZero, float tSub1, float time);
 	public RGB getAmbientColor(float ambience);
